backend/endpoints/requirements_dev.txt
wheel
aws-wsgi
flask
flask-cors
flask-sqlalchemy
ariadne
flask_alembic
invoke
psycopg2-binary
gitpython
